backend/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.7.0
build==1.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorclass==2.2.2
docopt==0.6.2
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-restx==1.2.0
Flask-SQLAlchemy==3.1.1
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jsonschema==4.17.3
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pip-tools==7.3.0
pip-upgrader==1.4.15
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyproject_hooks==1.0.0
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
six==1.16.0
SQLAlchemy==2.0.23
terminaltables==3.1.10
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==2.3.8
